+++ include/service/helper.h
@@ -18,11 +18,12 @@
 #pragma once
 
 #include <service/assert.h>
+#include <nul/types.h>
 
 /**
  * Check whether some address is in a range.
  */
-static inline bool in_range(unsigned long address, unsigned long base, unsigned long size) { return (base <= address) && (address <= base + size - 1); }
+static inline bool in_range(uintptr_t address, uintptr_t base, size_t size) { return (base <= address) && (address <= base + size - 1); }
 
 /**
  * Check whether X is true, output an error and return.
